requirements.txt
# Core dependencies
numpy>=1.19.0
pandas>=1.2.0
pyyaml>=5.4
click>=7.0
yfinance>=0.2.3

# Technical Analysis
TA-Lib>=0.4.24

# Data and Analysis
scikit-learn>=0.24.0
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=4.14.0

# Monitoring and utilities
psutil>=5.8.0
python-dotenv>=0.19.0

# Development
pytest>=6.2.0
black>=21.5b2
mypy>=0.910

# Visualization
streamlit>=1.2.0
dash>=2.0.0

# Validation
jsonschema>=3.2.0
